move fetcher out of core, and refactor fetcher to use separate tasks instead of a select loop
