+++ dim/src/core.rs
@@ -1,25 +1,15 @@
 use crate::logger::RequestLogger;
 use crate::routes;
 use crate::scanners;
+use crate::fetcher::PosterType;
 use crate::websocket;
 
 use once_cell::sync::OnceCell;
 
-use slog::debug;
-use slog::error;
 use slog::Logger;
 
-use tokio::sync::mpsc::unbounded_channel;
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::sync::mpsc::UnboundedSender;
-
-use std::collections::HashSet;
-use std::fs::File;
-use std::io::copy;
-use std::io::Cursor;
-use std::path::PathBuf;
-
-use self::fetcher::PosterType;
 
 use warp::Filter;
 
@@ -87,124 +77,6 @@ pub async fn run_scanners(log: Logger, tx: EventTx) {
                     .expect("Something went wrong with the fs-watcher");
             });
         }
-    }
-}
-
-pub mod fetcher {
-    use std::cmp::Ordering;
-    use std::collections::BTreeSet;
-    use std::time::Duration;
-
-    use super::*;
-
-    #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-    pub enum PosterType {
-        Banner(String),
-        Season(String),
-        Episode(String),
-    }
-
-    impl PartialOrd for PosterType {
-        fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-            Some(self.cmp(other))
-        }
-    }
-
-    impl Ord for PosterType {
-        fn cmp(&self, other: &Self) -> Ordering {
-            match (self, other) {
-                (PosterType::Banner(_), PosterType::Banner(_)) => Ordering::Equal,
-                (PosterType::Banner(_), PosterType::Season(_)) => Ordering::Greater,
-                (PosterType::Banner(_), PosterType::Episode(_)) => Ordering::Greater,
-                (PosterType::Season(_), PosterType::Banner(_)) => Ordering::Less,
-                (PosterType::Season(_), PosterType::Season(_)) => Ordering::Equal,
-                (PosterType::Season(_), PosterType::Episode(_)) => Ordering::Greater,
-                (PosterType::Episode(_), PosterType::Banner(_)) => Ordering::Less,
-                (PosterType::Episode(_), PosterType::Season(_)) => Ordering::Less,
-                (PosterType::Episode(_), PosterType::Episode(_)) => Ordering::Equal,
-            }
-        }
-    }
-
-    impl From<PosterType> for String {
-        fn from(poster: PosterType) -> Self {
-            match poster {
-                PosterType::Banner(st) | PosterType::Season(st) | PosterType::Episode(st) => st,
-            }
-        }
-    }
-
-    /// Function creates a task that fetches and caches posters from various sources.
-    pub async fn tmdb_poster_fetcher(log: Logger) {
-        let (tx, mut rx): (UnboundedSender<PosterType>, UnboundedReceiver<PosterType>) =
-            unbounded_channel();
-
-        let fut = async move {
-            // we need to cache which posters we've already fetched as the scanners arent generally
-            // aware of which assets are available or not.
-            let mut poster_cache = HashSet::<PosterType>::new();
-            let mut processing = BTreeSet::<PosterType>::new();
-
-            let mut timer = tokio::time::interval(Duration::from_millis(100));
-
-            loop {
-                tokio::select! {
-                    _ = timer.tick() => {
-                        while let Some(poster) = processing.pop_last() {
-                            let url: String = poster.clone().into();
-
-                            debug!(log, "Trying to cache {}", url);
-                            match reqwest::get(url.as_str()).await {
-                                Ok(resp) => {
-                                    if let Some(fname) = resp.url().path_segments().and_then(|segs| segs.last()) {
-                                        let meta_path = METADATA_PATH.get().unwrap();
-                                        let mut out_path = PathBuf::from(meta_path);
-                                        out_path.push(fname);
-
-                                        debug!(log, "Caching {} -> {:?}", url, out_path);
-
-                                        if let Ok(mut file) = File::create(out_path) {
-                                            if let Ok(bytes) = resp.bytes().await {
-                                                let mut content = Cursor::new(bytes);
-                                                if copy(&mut content, &mut file).is_ok() {
-                                                    continue;
-                                                }
-                                            }
-                                        }
-                                    }
-                                    error!(log, "Failed to cache {} locally, appending back into queue", &url);
-                                    processing.insert(poster);
-                                }
-                                Err(e) => {
-                                    error!(log, "Failed to cache {} locally, e={:?}", url, e);
-                                    processing.insert(poster);
-                                },
-                            }
-                        }
-
-                        assert!(processing.is_empty());
-                    }
-
-                    Some(poster) = rx.recv() => {
-                        if !poster_cache.contains(&poster) {
-                            debug!(log, "Inserting {:?} into queue", poster);
-                            processing.insert(poster.clone());
-                            poster_cache.insert(poster);
-                        }
-                    }
-                }
-
-                // NOTE(val): turns out that sometimes looping too fast can result in data
-                // magically disappearing. yield to be nice :)
-                tokio::task::yield_now().await;
-            }
-        };
-
-        tokio::spawn(fut);
-
-        METADATA_FETCHER_TX
-            .set(CloneOnDeref::new(tx))
-            .expect("Failed to set METADATA_FETCHER_TX");
     }
 }
