+++ dim/src/scanners/movie.rs
@@ -19,7 +19,8 @@ use slog::Logger;
 use events::Message;
 use events::PushEventType;
 
-use crate::core::{fetcher::PosterType, EventTx};
+use crate::core::EventTx;
+use crate::fetcher::PosterType;
 
 pub struct MovieMatcher<'a> {
     pub conn: &'a DbConnection,
